helpers: Support extensions with more than 1 dot.
Currenty file type detection based on extension fails, if there is a
filename of the form `mesh.1.off` instead of `mesh.off` (with only a
single dot). This fix replaces the custom code with the default
`os.path.splitext` function which always only returns the rightmost part
or an empty string.