Gmsh: Fix v2.2 reader producing bad cell data arrays
This bug shows up if all cells of the same type are not listed
consecutively in the file. This reader generates a new cell block each
time there is a cell type change in the file. Therefore, the cell data
array lengths should agree with the number of cells on each cell block.