Update test of gmsh4.1 reference file.
Need different reference values for the number of cells in the
CellBlocks and among the cell_sets, since the latter now also include
the bounding entities of each block (if available).

Also extended test to do a read-write check (this was not possible
before, since the gmsh writer could not deal with more than one cell
type).