Synchronize pyproject.toml with setup.cfg
This commit synchronizes the `pyproject.toml` file settings with the
amended version of `setup.cfg` that updates the `numpy` version
requirement. The rationale behind this synchronization is, again, to
make a `setuptools`-based installation reading (mostly) from
`setup.cfg` behave the same way as a `build`- or `flit`-based
installation.