Bump version in pyproject.toml, setup.cfg to 7.0.1
This commit bumps the project version in both `pyproject.toml` and
`setup.cfg` to 7.0.1, based on the last release tag in the `meshio`
upstream Git repository at `https://github.com/nschloe/meshio/issues`.

There is also a `meshio-rewrite` project on PyPI that seems to be at
version 7.0.6, but it is unclear to me whether that version is related
to the upstream `meshio` repository above, because the linked homepage
for that PyPI entry is inaccessible.